Deleting use-urls hook and it's usage in app.tsx file. Transferring all logic to urls.ts util https://github.com/SoftUni-Internal/exam-systems-issues/issues/781
