+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/App.tsx
@@ -27,7 +27,6 @@ import ParticipationsProvider from './hooks/use-participations';
 import ProblemsProvider from './hooks/use-problems';
 import SearchProvider from './hooks/use-search';
 import ServicesProvider from './hooks/use-services';
-import UrlsProvider from './hooks/use-urls';
 import UsersProvider from './hooks/use-users';
 import PageContent from './layout/content/PageContent';
 import PageFooter from './layout/footer/PageFooter';
@@ -41,7 +40,6 @@ const App = () => {
         UrlParamsProvider,
         RouteUrlParamsProvider,
         AppUrlsProvider,
-        UrlsProvider,
         ServicesProvider,
         PageProvider,
         LoadingProvider,
