+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/utils/urls.ts
@@ -1,44 +1,180 @@
-// TODO: https://github.com/SoftUni-Internal/exam-systems-issues/issues/133
+import isNil from 'lodash/isNil';
+
+import { SearchParams } from '../common/search-types';
+import {
+    IAllContestsUrlParams,
+    IDownloadProblemResourceUrlParams,
+    IDownloadSubmissionFileUrlParams,
+    IGetContestByProblemUrlParams,
+    IGetContestParticipationScoresForParticipantUrlParams,
+    IGetContestResultsParams,
+    IGetSearchResultsUrlParams,
+    IGetSubmissionDetailsByIdUrlParams,
+    IGetSubmissionResultsByProblemUrlParams,
+    IRetestSubmissionUrlParams,
+    IStartContestParticipationUrlParams,
+    ISubmitContestPasswordUrlParams,
+} from '../common/url-types';
+
 const
     {
         URLS:
-        {
-            UI_URL: baseUrl,
-            ADMINISTRATION_URL: administrationBaseUrl,
-        },
+            {
+                UI_URL: baseUrl,
+                ADMINISTRATION_URL: administrationBaseUrl,
+                PLATFORM_URL: platformBaseUrl,
+            },
     } = window;
 
-const loginSubmitUrl = `${baseUrl}/Account/Login`;
-const logoutUrl = `${baseUrl}/Account/Logout`;
+const baseApiUrl = `${baseUrl}/api`;
 
-const getProfileInfoUrl = `${baseUrl}/Users/GetProfileInfo`;
+// auth
+const getLoginSubmitUrl = () => `${baseUrl}/Account/Login`;
+const getLogoutUrl = () => `${baseUrl}/Account/Logout`;
 
-const getIndexContestsUrl = `${baseUrl}/Contests/GetForHomeIndex`;
-const startContestParticipationUrl = `${baseUrl}/Compete/Index/%id%?official=%official%`;
+const getUserAuthInfoUrl = () => `${baseApiUrl}/Users/GetUserAuthInfo`;
+const getPlatformRegisterUrl = () => `${platformBaseUrl}/identity/register`;
 
-const getSubmissionDetailsUrl = `${baseUrl}/Submissions/Details`;
-const getSubmissionsForProfileUrl = `${baseUrl}/Submissions/GetForProfile`;
-const submitUrl = `${baseUrl}/Compete/Submit`;
+// admin
+const getAdministrationContestsGridUrl = () => `${administrationBaseUrl}/Contests`;
+const getAdministrationNavigation = () => '/administration';
+const getAdministrationRetestSubmission = ({ id }: IRetestSubmissionUrlParams) => `
+${administrationBaseUrl}/Submissions/Retest?PK=${id}`;
 
-const getParticipationsForProfileUrl = `${baseUrl}/Participations/GetForProfile`;
+// profile
+const getProfileInfoUrl = () => `${baseApiUrl}/Users/GetProfileInfo`;
+const getSubmissionsForProfileUrl = () => `${baseApiUrl}/Submissions/GetForProfile`;
+const getParticipationsForProfileUrl = () => `${baseApiUrl}/Participations/GetForProfile`;
 
-const downloadProblemResourceUrl = `${baseUrl}/ProblemResources/GetResource/%id%`;
+// contests
+const getIndexContestsUrl = () => `${baseApiUrl}/Contests/GetForHomeIndex`;
+const getAllContestsUrl = ({ filters, sorting, page }: IAllContestsUrlParams) => {
+    const filtersQuery = `${filters
+        .map(({ value, type }) => `${type.toLowerCase()}=${value}`)
+        .join('&')
+    }`;
 
-const getSubmissionResultsByProblem = `${baseUrl}/Submissions/GetSubmissionResultsByProblem/%id%?isOfficial=%isOfficial%&take=%take%`;
+    const sortingQuery = `${sorting
+        .map(({ value, type }) => `${type.toLowerCase()}=${value}`)
+        .join('&')
+    }`;
 
-const administrationContestsGridUrl = `${administrationBaseUrl}/Contests`;
+    const pageQuery = isNil(page)
+        ? ''
+        : `page=${page}`;
+
+    return `${baseApiUrl}/Contests/GetAll?${filtersQuery}&${sortingQuery}&${pageQuery}`;
+};
+
+const getRegisterForContestUrl = ({
+    id,
+    isOfficial,
+}: IStartContestParticipationUrlParams) => `${baseApiUrl}/Contests/Register/${id}?official=${isOfficial}`;
+
+const getSubmitContestPasswordUrl = ({
+    id,
+    isOfficial,
+}: ISubmitContestPasswordUrlParams) => `${baseApiUrl}/Contests/SubmitContestPassword/${id}?official=${isOfficial}`;
+
+const getStartContestParticipationUrl = ({
+    id,
+    isOfficial,
+}: IStartContestParticipationUrlParams) => `${baseApiUrl}/Compete/Index/${id}?official=${isOfficial}`;
+
+const getContestParticipantScoresForParticipantUrl =
+    ({ participantId }: IGetContestParticipationScoresForParticipantUrlParams) => `
+    ${baseApiUrl}/ParticipantScores/GetScoresForParticipant/${participantId}`;
+
+const getContestByProblemUrl = ({ problemId }: IGetContestByProblemUrlParams) => `${baseApiUrl}/Contests/GetByProblem/${problemId}`;
+
+const getCategoriesTreeUrl =
+    () => `${baseApiUrl}/ContestCategories/GetCategoriesTree`;
+
+const getContestResultsUrl = ({
+    id,
+    official,
+    full,
+} : IGetContestResultsParams) => `${baseApiUrl}/ContestResults/GetResults/${id}?official=${official}&full=${full}`;
+
+// submissions
+const getSubmissionResultsByProblemUrl = ({
+    problemId,
+    isOfficial,
+    take,
+}: IGetSubmissionResultsByProblemUrlParams) => `
+    ${baseApiUrl}/Submissions/GetSubmissionResultsByProblem/${problemId}?isOfficial=${isOfficial}&take=${take}`;
+
+const getSubmissionDetailsResultsUrl = ({
+    submissionId,
+    isOfficial,
+    take,
+}: IGetSubmissionDetailsByIdUrlParams) => `
+    ${baseApiUrl}/Submissions/GetSubmissionDetailsResults/${submissionId}?isOfficial=${isOfficial}&take=${take}`;
+const getPublicSubmissionsUrl = () => `${baseApiUrl}/Submissions/Public`;
+const getSubmissionsTotalCountUrl = () => `${baseApiUrl}/Submissions/TotalCount`;
+const getSubmissionsDetailsUrl = () => `${baseApiUrl}/Submissions/Details`;
+const getSubmissionDetailsByIdUrl =
+    ({ submissionId }: IGetSubmissionDetailsByIdUrlParams) => `${getSubmissionsDetailsUrl()}/${submissionId}`;
+const getSubmitUrl = () => `${baseApiUrl}/Compete/Submit`;
+const getSubmitFileUrl = () => `${baseApiUrl}/Compete/SubmitFileSubmission`;
+const getSubmissionFileDownloadUrl =
+    ({ id }: IDownloadSubmissionFileUrlParams) => `${baseApiUrl}/Submissions/Download/${id}`;
+
+// Submission types
+const getAllContestStrategyFiltersUrl =
+    () => `${baseApiUrl}/SubmissionTypes/GetAllOrderedByLatestUsage`;
+
+// problem resources
+const getDownloadProblemResourceUrl = ({ id }: IDownloadProblemResourceUrlParams) => `${baseApiUrl}/ProblemResources/GetResource/${id}`;
+
+// Statistics
+const getHomeStatisticsUrl = () => `${baseApiUrl}/StatisticsPreview/GetForHome`;
+
+// Search
+const getSearchResults = ({ searchTerm, page, selectedTerms }: IGetSearchResultsUrlParams) => {
+    const searchQuery = `${SearchParams.search}=${searchTerm}`;
+
+    const pageQuery = `page=${page}`;
+
+    const selectedTermQuery = `&${selectedTerms
+        .map(({ key, value }) => `${key}=${value}`)
+        .join('&')
+    }`;
+
+    return `${baseApiUrl}/Search/GetSearchResults?${searchQuery}${selectedTermQuery}&${pageQuery}`;
+};
 
 export {
-    loginSubmitUrl,
-    logoutUrl,
+    getLoginSubmitUrl,
+    getLogoutUrl,
+    getUserAuthInfoUrl,
+    getPlatformRegisterUrl,
+    getAdministrationContestsGridUrl,
+    getAdministrationNavigation,
+    getAdministrationRetestSubmission,
     getProfileInfoUrl,
-    getIndexContestsUrl,
-    startContestParticipationUrl,
-    getSubmissionDetailsUrl,
     getSubmissionsForProfileUrl,
     getParticipationsForProfileUrl,
-    downloadProblemResourceUrl,
-    getSubmissionResultsByProblem,
-    submitUrl,
-    administrationContestsGridUrl,
+    getIndexContestsUrl,
+    getAllContestsUrl,
+    getSubmitContestPasswordUrl,
+    getRegisterForContestUrl,
+    getStartContestParticipationUrl,
+    getContestParticipantScoresForParticipantUrl,
+    getContestResultsUrl,
+    getCategoriesTreeUrl,
+    getContestByProblemUrl,
+    getSubmissionResultsByProblemUrl,
+    getSubmissionDetailsResultsUrl,
+    getPublicSubmissionsUrl,
+    getSubmissionsTotalCountUrl,
+    getSubmissionsDetailsUrl,
+    getSubmissionDetailsByIdUrl,
+    getSubmitUrl,
+    getSubmitFileUrl,
+    getSubmissionFileDownloadUrl,
+    getAllContestStrategyFiltersUrl,
+    getDownloadProblemResourceUrl,
+    getHomeStatisticsUrl,
+    getSearchResults,
 };
